// web/default/src/features/system-logs/types.ts
/*
Copyright (C) 2023-2026 QuantumNous

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License
along with this program. If not, see <https://www.gnu.org/licenses/>.

For commercial licensing, please contact support@quantumnous.com
*/
/**
 * Type definitions for the system logs feature
 */

export interface SystemLog {
  id: number
  user_id: number
  username: string
  created_at: number
  type: number
  content: string
  quota: number
  channel: number
  channel_name: string
  token_id: number
  token_name: string
  model_name: string
  group: string
  ip: string
  user_agent: string
  request_id?: string
  other: string
}

export interface SystemLogFilters {
  startTime?: Date
  endTime?: Date
  username?: string
  ip?: string
  group?: string
  modelName?: string
}

export interface GetSystemLogsParams {
  p?: number
  page_size?: number
  type?: number
  username?: string
  model_name?: string
  ip?: string
  group?: string
  start_timestamp?: number
  end_timestamp?: number
}

export interface GetSystemLogsResponse {
  success: boolean
  message?: string
  data?: {
    items: SystemLog[]
    total: number
    page: number
    page_size: number
  }
}